+++ Tests/SentryTests/SentryHubTests.swift
@@ -27,7 +27,7 @@ class SentryHubTests: XCTestCase {
         let random = TestRandom(value: 0.5)
         let queue = DispatchQueue(label: "SentryHubTests", qos: .utility, attributes: [.concurrent])
         let dispatchQueueWrapper = TestSentryDispatchQueueWrapper()
-        
+
         init() {
             options = Options()
             options.dsn = SentryHubTests.dsnAsString
@@ -755,24 +755,28 @@ class SentryHubTests: XCTestCase {
             XCTAssertEqual(fixture.scope, errorArguments.scope)
         }
     }
-    
-    func testCaptureErrorWithSessionWithScope() {
+
+    func testCaptureErrorWithSessionWithScope() throws {
+        // Arrange
         let sut = fixture.getSut()
         sut.startSession()
+
+        // Act
         sut.capture(error: fixture.error, scope: fixture.scope).assertIsNotEmpty()
-        
-        XCTAssertEqual(1, fixture.client.captureErrorWithSessionInvocations.count)
-        if let errorArguments = fixture.client.captureErrorWithSessionInvocations.first {
+
+        // Assert
+        XCTAssertEqual(1, fixture.client.captureErrorWithScopeInvocations.count)
+        if let errorArguments = fixture.client.captureErrorWithScopeInvocations.first {
             XCTAssertEqual(fixture.error, errorArguments.error as NSError)
-            
-            XCTAssertEqual(1, errorArguments.session?.errors)
-            XCTAssertEqual(SentrySessionStatus.ok, errorArguments.session?.status)
-            
             XCTAssertEqual(fixture.scope, errorArguments.scope)
         }
         
         // only session init is sent
-        XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
+        XCTAssertEqual(fixture.client.captureSessionInvocations.count, 1)
+
+        let actualSession = try XCTUnwrap(sut.session)
+        XCTAssertEqual(actualSession.errors, 1)
+        XCTAssertEqual(actualSession.status, SentrySessionStatus.ok)
     }
     
     func testCaptureErrorBeforeSessionStart() {
@@ -827,23 +831,6 @@ class SentryHubTests: XCTestCase {
         }
     }
     
-    func testCaptureWithoutIncreasingErrorCount() {
-        let sut = fixture.getSut()
-        sut.startSession()
-        fixture.client.callSessionBlockWithIncrementSessionErrors = false
-        sut.capture(error: fixture.error, scope: fixture.scope).assertIsNotEmpty()
-        
-        XCTAssertEqual(1, fixture.client.captureErrorWithSessionInvocations.count)
-        if let errorArguments = fixture.client.captureErrorWithSessionInvocations.first {
-            XCTAssertEqual(fixture.error, errorArguments.error as NSError)
-            XCTAssertNil(errorArguments.session)
-            XCTAssertEqual(fixture.scope, errorArguments.scope)
-        }
-        
-        // only session init is sent
-        XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
-    }
-    
     func testCaptureErrorWithoutScope() {
         fixture.getSut(fixture.options, fixture.scope).capture(error: fixture.error).assertIsNotEmpty()
         
@@ -873,77 +860,81 @@ class SentryHubTests: XCTestCase {
             XCTAssertEqual(fixture.scope, errorArguments.scope)
         }
     }
-    
-    func testCaptureExceptionWithSessionWithScope() {
+
+    func testCaptureMultipleExceptionsInParallel_IncrementsSessionCount() throws {
+        // Arrange
+        let captureCount = 100
+
+        let sut = fixture.getSut()
+
+        sut.startSession()
+
+        let expectation = XCTestExpectation(description: "Capture error")
+        expectation.expectedFulfillmentCount = captureCount
+        expectation.assertForOverFulfill = true
+
+        // Act
+        for _ in 0..<captureCount {
+
+            fixture.queue.async {
+                sut.capture(exception: self.fixture.exception, scope: self.fixture.scope)
+                expectation.fulfill()
+            }
+        }
+
+        wait(for: [expectation], timeout: 5.0)
+
+        // Assert
+        let session = try XCTUnwrap(sut.session)
+        XCTAssertEqual(session.errors, UInt(captureCount))
+    }
+
+    func testCaptureMultipleErrorsInParallel_IncrementsSessionCount() throws {
+        // Arrange
+        let captureCount = 100
+
+        let sut = fixture.getSut()
+
+        sut.startSession()
+
+        let expectation = XCTestExpectation(description: "Capture error")
+        expectation.expectedFulfillmentCount = captureCount
+        expectation.assertForOverFulfill = true
+
+        // Act
+        for _ in 0..<captureCount {
+
+            fixture.queue.async {
+                sut.capture(error: self.fixture.error, scope: self.fixture.scope)
+                expectation.fulfill()
+            }
+        }
+
+        wait(for: [expectation], timeout: 5.0)
+
+        // Assert
+        let session = try XCTUnwrap(sut.session)
+        XCTAssertEqual(session.errors, UInt(captureCount))
+    }
+
+    func testCaptureEventIncrementingSessionErrorCount_WithStartedSession_OnlySendsSessionInit() {
         let sut = fixture.getSut()
         sut.startSession()
-        sut.capture(exception: fixture.exception, scope: fixture.scope).assertIsNotEmpty()
-        
-        XCTAssertEqual(1, fixture.client.captureExceptionWithSessionInvocations.count)
-        if let exceptionArguments = fixture.client.captureExceptionWithSessionInvocations.first {
-            XCTAssertEqual(fixture.exception, exceptionArguments.exception)
-            
-            XCTAssertEqual(1, exceptionArguments.session?.errors)
-            XCTAssertEqual(SentrySessionStatus.ok, exceptionArguments.session?.status)
-            
-            XCTAssertEqual(fixture.scope, exceptionArguments.scope)
+
+        let event = fixture.event
+        sut.captureErrorEvent(event: event).assertIsNotEmpty()
+
+        XCTAssertEqual(fixture.client.captureEventIncrementingSessionErrorCountInvocations.count, 1)
+        if let eventArguments = fixture.client.captureEventIncrementingSessionErrorCountInvocations.first {
+            XCTAssertEqual(eventArguments.event.eventId, event.eventId)
+
+            XCTAssertEqual(eventArguments.scope, sut.scope)
         }
-        
+
         // only session init is sent
         XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
     }
-    
-    func testCaptureExceptionWithoutIncreasingErrorCount() {
-        let sut = fixture.getSut()
-        sut.startSession()
-        fixture.client.callSessionBlockWithIncrementSessionErrors = false
-        sut.capture(exception: fixture.exception, scope: fixture.scope).assertIsNotEmpty()
-        
-        XCTAssertEqual(1, fixture.client.captureExceptionWithSessionInvocations.count)
-        if let exceptionArguments = fixture.client.captureExceptionWithSessionInvocations.first {
-            XCTAssertEqual(fixture.exception, exceptionArguments.exception)
-            XCTAssertNil(exceptionArguments.session)
-            XCTAssertEqual(fixture.scope, exceptionArguments.scope)
-        }
-        
-        // only session init is sent
-        XCTAssertEqual(1, fixture.client.captureSessionInvocations.count)
-    }
-    
-    func testCaptureMultipleExceptionWithSessionInParallel() {
-        let captureCount = 100
-        captureConcurrentWithSession(count: captureCount) { sut in
-            sut.capture(exception: self.fixture.exception, scope: self.fixture.scope)
-        }
-        
-        let invocations = fixture.client.captureExceptionWithSessionInvocations.invocations
-        XCTAssertEqual(captureCount, invocations.count)
-        for i in 1...captureCount {
-            // The session error count must not be in order as we use a concurrent DispatchQueue
-            XCTAssertTrue(
-                invocations.contains { $0.session!.errors == i },
-                "No session captured with \(i) amount of errors."
-            )
-        }
-    }
-    
-    func testCaptureMultipleErrorsWithSessionInParallel() {
-        let captureCount = 100
-        captureConcurrentWithSession(count: captureCount) { sut in
-            sut.capture(error: self.fixture.error, scope: self.fixture.scope)
-        }
-        
-        let invocations = fixture.client.captureErrorWithSessionInvocations.invocations
-        XCTAssertEqual(captureCount, invocations.count)
-        for i in 1..<captureCount {
-            // The session error count must not be in order as we use a concurrent DispatchQueue
-            XCTAssertTrue(
-                invocations.contains { $0.session!.errors == i },
-                "No session captured with \(i) amount of errors."
-            )
-        }
-    }
-    
+
     func testCaptureClientIsNil_ReturnsEmptySentryId() {
         sut.bindClient(nil)
         
@@ -1333,26 +1324,6 @@ class SentryHubTests: XCTestCase {
         })
     }
     
-    private func captureConcurrentWithSession(count: Int, _ capture: @escaping (SentryHubInternal) -> Void) {
-        let sut = fixture.getSut()
-        sut.startSession()
-        
-        let queue = fixture.queue
-
-        let expectation = XCTestExpectation(description: "Capture should be called \(count) times")
-        expectation.expectedFulfillmentCount = count
-
-        for _ in 0..<count {
-
-            queue.async {
-                capture(sut)
-                expectation.fulfill()
-            }
-        }
-
-        wait(for: [expectation], timeout: 5.0)
-    }
-    
     func testModifyIntegrationsConcurrently() {
         
         let sut = fixture.getSut()
@@ -1464,6 +1435,33 @@ class SentryHubTests: XCTestCase {
                                                                  ["mechanism": ["other-key": false]]]
                                                             ]
                                                          ]))
+    }
+
+    func testBindClient_SetsSessionDelegate() throws {
+        // Arrange
+        let sut = fixture.getSut()
+        let currentClient = fixture.client
+        let newClient = SentryClientInternal(options: fixture.options)
+
+        // Act
+        sut.bindClient(newClient)
+
+        // Assert
+        XCTAssertNil(currentClient.sessionDelegate)
+        // We only assert if it's not nil for a safety check. Other unit tests validate the correctness of the delegate methods.
+        XCTAssertNotNil(newClient?.sessionDelegate)
+    }
+
+    func testBindNilClient_SetsSessionDelegateToNil() {
+        // Arrange
+        let sut = fixture.getSut()
+        let client = fixture.client
+
+        // Act
+        sut.bindClient(nil)
+
+        // Assert
+        XCTAssertNil(client.sessionDelegate)
     }
 
     private func captureEventEnvelope(level: SentryLevel) {